Switch from proxy connect to local proxy
