+++ internal/command/local/connect.go
@@ -74,12 +74,6 @@ func runConnect(cmd *cobra.Command, out io.Writer, cfg *config.LocalConnect, arg
 		connConfig.KubeConfigPath = &kcp
 	}
 
-	// Get control-plane proxy url
-	proxyURL, err := config.GetProxyURL()
-	if err != nil {
-		return err
-	}
-
 	// compute ConnectInvocationConfig
 	ciConfig := &config.ConnectInvocationConfig{
 		WithRootManager: !cfg.Unprivileged,
@@ -95,8 +89,8 @@ func runConnect(cmd *cobra.Command, out io.Writer, cfg *config.LocalConnect, arg
 			Username: user.Username,
 		},
 		ConnectionConfig: connConfig,
-		ProxyURL:         proxyURL,
-		APIKey:           config.GetAPIKey(),
+		ProxyURL:         cfg.ProxyURL,
+		APIKey:           cfg.GetAPIKey(),
 		Debug:            cfg.LocalConfig.Debug,
 	}
 	if cfg.DumpCIConfig {
