+++ internal/config/local.go
@@ -1,9 +1,13 @@
 package config
 
 import (
+	"bytes"
+	"errors"
 	"flag"
 	"fmt"
+	"net/url"
 	"os"
+	"regexp"
 	"strconv"
 	"time"
 
@@ -20,12 +24,24 @@ const (
 type Local struct {
 	*API
 
+	ProxyURL string
 	// initialized from ~/.signadot/config.yaml
 	LocalConfig *config.Config
 }
 
+func (l *Local) InitLocalProxyConfig() error {
+	var err error
+	if err = l.API.InitAPIConfig(); err != nil {
+		return err
+	}
+	if l.ProxyURL, err = l.GetProxyURL(); err != nil {
+		return err
+	}
+	return nil
+}
+
 func (l *Local) InitLocalConfig() error {
-	if err := l.API.InitAPIConfig(); err != nil {
+	if err := l.InitLocalProxyConfig(); err != nil {
 		return err
 	}
 
@@ -78,6 +94,22 @@ func (l *Local) GetConnectionConfig(cluster string) (*config.ConnectionConfig, e
 		}
 	}
 	return nil, fmt.Errorf("no such cluster %q, expecting one of %v", cluster, clusters)
+}
+
+func (l *Local) GetProxyURL() (string, error) {
+	// Allow Proxy URL to be overridden (e.g. for talking to dev/staging).
+	if proxyURL := viper.GetString("proxy_url"); proxyURL != "" {
+		_, err := url.Parse(proxyURL)
+		if err != nil {
+			return "", fmt.Errorf("invalid proxy_url: %w", err)
+		}
+		return proxyURL, nil
+	}
+	return "https://proxy.signadot.com", nil
+}
+
+func (l *Local) GetAPIKey() string {
+	return viper.GetString("api_key")
 }
 
 type LocalConnect struct {
@@ -157,4 +189,111 @@ type LocalStatus struct {
 
 func (c *LocalStatus) AddFlags(cmd *cobra.Command) {
 	cmd.Flags().BoolVar(&c.Details, "details", false, "display status details")
+}
+
+type LocalProxy struct {
+	*Local
+
+	// Flags
+	Sandbox       string
+	RouteGroup    string
+	Cluster       string
+	ProxyMappings []ProxyMapping
+}
+
+func (lp *LocalProxy) Validate() error {
+	c := 0
+	if lp.Sandbox != "" {
+		c += 1
+	}
+	if lp.RouteGroup != "" {
+		c += 1
+	}
+	if lp.Cluster != "" {
+		c += 1
+	}
+
+	if c == 0 {
+		return errors.New("you should specify one of '--sandbox', '--routegroup' or '--cluster'")
+	}
+	if c > 1 {
+		return errors.New("only one of '--sandbox', '--routegroup' or '--cluster' should be specified")
+	}
+
+	for i := range lp.ProxyMappings {
+		pm := &lp.ProxyMappings[i]
+		if err := pm.Validate(); err != nil {
+			return err
+		}
+
+	}
+	return nil
+}
+
+type ProxyMapping struct {
+	TargetProto string
+	TargetAddr  string
+	BindAddr    string
+}
+
+func (pm *ProxyMapping) Validate() error {
+	switch pm.TargetProto {
+	case "tcp":
+	case "http":
+	case "https":
+	case "grpc":
+	default:
+		return fmt.Errorf("unsupported '%s' protocol, only 'tcp', 'http', 'https' and 'grpc' are supported", pm.TargetProto)
+	}
+	return nil
+}
+
+func (pm *ProxyMapping) GetTarget() string {
+	return fmt.Sprintf("%s://%s", pm.TargetProto, pm.TargetAddr)
+}
+
+func (pm *ProxyMapping) String() string {
+	return fmt.Sprintf("%s://%s|%s", pm.TargetProto, pm.TargetAddr, pm.BindAddr)
+}
+
+type proxyMappings []ProxyMapping
+
+func (pms *proxyMappings) String() string {
+	b := bytes.NewBuffer(nil)
+	for i := range *pms {
+		pm := &(*pms)[i]
+		if i != 0 {
+			fmt.Fprintf(b, " ")
+		}
+		fmt.Fprintf(b, "--map %s", pm)
+	}
+	return b.String()
+}
+
+// Set appends a new argument  to instance of Nargs
+func (pms *proxyMappings) Set(arg string) error {
+	regex := regexp.MustCompile(`^(.+?)://(.+?)\|(.+)$`)
+	matches := regex.FindStringSubmatch(arg)
+	if matches == nil || len(matches) != 4 {
+		return fmt.Errorf("invalid format, expected \"<target-protocol>://<target-addr>|<bind-addr>\"")
+	}
+
+	*pms = append(*pms, ProxyMapping{
+		TargetProto: matches[1],
+		TargetAddr:  matches[2],
+		BindAddr:    matches[3],
+	})
+	return nil
+}
+
+// Type is a no-op
+func (pms *proxyMappings) Type() string {
+	return "string"
+}
+
+func (lp *LocalProxy) AddFlags(cmd *cobra.Command) {
+	cmd.Flags().StringVarP(&lp.Sandbox, "sandbox", "s", "", "run the proxy in the context of the specificed sandbox")
+	cmd.Flags().StringVarP(&lp.RouteGroup, "routegroup", "r", "", "run the proxy in the context of the specificed routegroup")
+	cmd.Flags().StringVarP(&lp.Cluster, "cluster", "c", "", "target cluster")
+	cmd.Flags().VarP((*proxyMappings)(&lp.ProxyMappings), "map", "m", "--map <target-protocol>://<target-addr>|<bind-addr>")
 }
